refactor(slider-web,range-slider-web): move shared formatting logic to widget-plugin-platform
createValueFormatter and createMarks extracted to widget-plugin-platform/utils so both
slider-web and range-slider-web consume a single implementation without duplication.
Tests moved to platform; duplicate tests removed from widget packages.
Range-slider-web now uses a single formatter derived from lowerBoundAttribute.
