+++ packages/pluggableWidgets/range-slider-web/src/utils/helpers.ts
@@ -1,17 +1,1 @@
-import { Big } from "big.js";
-import { NumberFormatter } from "mendix";
-
-export type ValueFormatter = (value: number) => string;
-
-/**
- * Builds a value formatter from the attribute's own Mendix NumberFormatter, overriding only the
- * decimal precision. Reusing the runtime formatter means the decimal separator and thousands
- * grouping follow the user's session locale and the attribute's `groupDigits` setting automatically.
- */
-export function createValueFormatter(formatter: NumberFormatter, decimalPlaces: number): ValueFormatter {
-    const configured = formatter.withConfig({
-        groupDigits: formatter.config.groupDigits,
-        decimalPrecision: decimalPlaces
-    });
-    return (value: number) => configured.format(new Big(value));
-}
+export { createValueFormatter, type ValueFormatter } from "@mendix/widget-plugin-platform/utils/number-formatter";
