+++ packages/pluggableWidgets/slider-web/package.json
@@ -44,6 +44,7 @@
     },
     "dependencies": {
         "@mendix/widget-plugin-component-kit": "workspace:*",
+        "@mendix/widget-plugin-platform": "workspace:*",
         "@rc-component/slider": "^1.0.1",
         "@rc-component/tooltip": "^1.3.3",
         "classnames": "^2.5.1"
@@ -55,7 +56,6 @@
         "@mendix/prettier-config-web-widgets": "workspace:*",
         "@mendix/run-e2e": "workspace:*",
         "@mendix/widget-plugin-hooks": "workspace:*",
-        "@mendix/widget-plugin-platform": "workspace:*",
         "cross-env": "^7.0.3"
     }
 }
