+++ packages/pluggableWidgets/slider-web/src/Slider.editorPreview.tsx
@@ -17,7 +17,7 @@ export function preview(props: SliderPreviewProps): ReactNode {
         max: values.max,
         numberOfMarks: props.noOfMarkers ?? 2,
         decimalPlaces,
-        format: (value: number) => value.toFixed(decimalPlaces)
+        format: (v: number) => v.toFixed(decimalPlaces)
     });
     const style = getStyleProp({
         orientation: props.orientation,
