+++ packages/pluggableWidgets/slider-web/src/utils/__tests__/createHandleRender.spec.tsx
@@ -13,7 +13,6 @@ const defaultRenderProps = {
 
 const mockNode = <div />;
 
-// Deterministic stand-in for createValueFormatter's output.
 const formatWith =
     (decimalPlaces: number, decimalSeparator = ".") =>
     (value: number): string => {
