+++ packages/shared/widget-plugin-platform/jest.config.cjs
@@ -15,6 +15,7 @@ module.exports = {
         ]
     },
     moduleNameMapper: {
+        "^big\\.js$": "<rootDir>/../../../node_modules/big.js/big.js",
         "(.+)\\.js": "$1"
     },
     extensionsToTreatAsEsm: [".ts"],
