+++ packages/shared/widget-plugin-platform/package.json
@@ -36,6 +36,7 @@
         "@mendix/tsconfig-web-widgets": "workspace:*",
         "@swc/core": "^1.7.26",
         "@swc/jest": "^0.2.36",
+        "big.js": "^6.2.1",
         "classnames": "^2.5.1",
         "jest-environment-jsdom": "^29.7.0"
     }
